NSQConsumer: shutdown message processing on close
Executor does not have a shutdown method, only ExecutorService.

1. Shutdown executor first so no new message processing can start.
2. Shutdown scheduler
3. Tell NSQD we are closing down so stop sending messages.

Not 100% sure this is the desired order in a shutdown. Any suggestions
would be nice.

A caveat here is that new tasks submitted to executor will be rejected
after executor is shutdown. Maybe this is not desirable. At least it
will give the opportunity to specifically reject messages queued after
shutdown has been started.